fixed "index out of bounds" error
fixed "index out of bounds" error -- still double check though

problem was line 44: if the sentence length was <6, for example 3, it couldnt go range(6,  3), so error. also, if  length is 6, will go range(6, 6), which leads to a empty inputNumbers, also giving an error.

fix: if <6, make it go range( length-1, length). if =6, make it go range(3, 6)

@Leahie -- line36 -- !CHECK: @LEAH --> not exactly sure if numBlanks is actually the number of blanks; double check porfa